mini_projects/csilla.toth/meeting_minutes_cli_hazi3/requirements.txt
requests
openai>=1.0.0
python-dotenv>=1.0.0
